Extract test helpers from src/ to tests/
