+++ tests/unit/helpers.ts
@@ -1,0 +1,10 @@
+import {RuleTester} from 'eslint';
+
+export function trimTestCases<
+  T extends RuleTester.InvalidTestCase | RuleTester.ValidTestCase
+>(testCase: T): T {
+  return {
+    ...testCase,
+    code: testCase.code.trim()
+  };
+}
